refactor: create lib/ folder and move dispatches, newsfeed, intervals and sync-from-api into it
