+++ src/events/ready/start-intervals.js
@@ -1,4 +1,4 @@
-const { startIntervals } = require('../../utils/intervals');
+const { startIntervals } = require('../../lib/intervals');
 
 module.exports = (client) => {
   startIntervals(client);
